Handle declaring of results & add button to start game
Declaring results is handled separately now. Also, a button has been
introduced, clicking on which starts the game. It also fixes the bug where
if the user didn't open console before the game began, they will not be
able to see any progress of game in the console, instead they will be able
to see the entire gameplay at once when the game ends.