Design improvement & bug fixes related to styling
There are no more "_" instead they are replaced with empty strings, so the
gameboard looks neat now. Also, there was this bug that whenever you
clicked already filled cell, although it was throwing the error and the
value was not getting registered, but the color of the cell was changing.
This was due to the fact that class responsible for color change was being
added earlier than checking for valid empty cell. It is fixed now.