Add restart game button & improve start game button
Start game button was causing lots of issues earlier. It was supposed to
be pressed only one time when the page loads, if a user presses it after
the game was started, it was troublesome. To counter this, now it gets
disabled once the game starts. Also, restart button has been implemented
which gives the user a fresh start.